uclibc: Fix double promotion warning
Add casts where necessary to convince clang that the promotion of float
to double is intentional.

+++ libm/nan.c
@@ -28,7 +28,7 @@ double nan (const char *tagp)
 		sprintf (buf, "NAN(%s)", tagp);
 		return strtod (buf, NULL);
 	}
-	return NAN;
+	return (double)NAN;
 }
 libm_hidden_def(nan)
 
@@ -53,7 +53,7 @@ long double nanl (const char *tagp)
 		sprintf (buf, "NAN(%s)", tagp);
 		return strtold (buf, NULL);
 	}
-	return NAN;
+	return (long double)NAN;
 }
 libm_hidden_def(nanl)
 #endif
